Convert BsSpinner to typescript
