+++ ember-bootstrap/addon/components/bs-spinner.hbs
@@ -1,9 +1,17 @@
-{{! @glint-nocheck }}
 <div
-  class="{{this.spinnerClass}} {{if @size (bs-size-class this.spinnerClass @size)}} {{bs-type-class "text" @type default="primary" outline=false}}"
+  class="{{this.spinnerClass}}
+    {{if @size (bs-size-class this.spinnerClass @size)}}
+    {{bs-type-class 'text' @type default="primary" outline=false}}"
   role="status"
-  ...attributes>
+  ...attributes
+>
   {{#if (has-block)}}
-    <span class={{if (macroCondition (macroGetOwnConfig "isBS5")) "visually-hidden" "sr-only"}}>{{yield}}</span>
+    <span
+      class={{if
+        (macroCondition (macroGetOwnConfig "isBS5"))
+        "visually-hidden"
+        "sr-only"
+      }}
+    >{{yield}}</span>
   {{/if}}
 </div>
