+++ ember-bootstrap/addon/template-registry.ts
@@ -1,7 +1,9 @@
 import type BsButton from './components/bs-button';
 import type BsCollapse from './components/bs-collapse';
+import type BsSpinner from './components/bs-spinner';
 
 export default interface EmberBootstrapRegistry {
   BsButton: typeof BsButton;
   BsCollapse: typeof BsCollapse;
+  BsSpinner: typeof BsSpinner;
 }
