+++ test-app/tests/integration/components/bs-spinner-test.ts
@@ -1,8 +1,10 @@
-import { module, test } from 'qunit';
+import { testForBootstrap } from '../../helpers/bootstrap';
+
+import { module } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
-import { testForBootstrap } from '../../helpers/bootstrap';
+import { test } from '../../helpers/bootstrap';
 
 module('Integration | Component | bs-spinner', function (hooks) {
   setupRenderingTest(hooks);
@@ -84,9 +86,9 @@ module('Integration | Component | bs-spinner', function (hooks) {
   test('spinner has HTML attributes', async function (assert) {
     await render(hbs`<BsSpinner id='test' />`);
 
-    assert.equal(this.element.querySelector('div').getAttribute('id'), 'test');
+    assert.equal(this.element.querySelector('div')?.getAttribute('id'), 'test');
     assert.equal(
-      this.element.querySelector('div').getAttribute('role'),
+      this.element.querySelector('div')?.getAttribute('role'),
       'status',
     );
   });
